Added important note to conf.d file
Thank you for the hard work on this project! I understand it's in maintenance mode. This is a pure documentation fix, so it shouldn't be too big of an ask, I hope.

I thought it would have been reasonable to let ssl_certificate_key be named anything the user desires.
Unfortunately, according to `scripts/util.sh:parse_domains()` it needs to be called `privkey.pem`, otherwise it won't work.
This wasted a lot of my time, would like to add a simple comment to prevent others from falling into the same trap.

Thank you.